wip - trainable move out of ops.hpp
